Enum nextest_filtering::errors::ParseSingleError

#[non_exhaustive]
pub enum ParseSingleError {

    InvalidRegex {
        span: SourceSpan,
        message: String,
    },
    InvalidGlob {
        span: SourceSpan,
        error: GlobConstructError,
    },
    InvalidRegexWithoutMessage(SourceSpan),
    ExpectedCloseRegex(SourceSpan),
    InvalidOrOperator(SourceSpan),
    InvalidAndOperator(SourceSpan),
    UnexpectedArgument(SourceSpan),
    UnexpectedComma(SourceSpan),
    InvalidString(SourceSpan),
    ExpectedOpenParenthesis(SourceSpan),
    ExpectedCloseParenthesis(SourceSpan),
    InvalidEscapeCharacter(SourceSpan),
    ExpectedExpr(SourceSpan),
    ExpectedEndOfExpression(SourceSpan),
    NoPackageMatch(SourceSpan),
    InvalidPlatformArgument(SourceSpan),
    Unknown,

}

An individual error that occurred while parsing a filter expression.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

InvalidRegex

Fields

span: SourceSpan

The part of the input that failed.

message: String

A message indicating the failure.

An invalid regex was encountered.

InvalidGlob

Fields

span: SourceSpan

The part of the input that failed.

error: GlobConstructError

The underlying error.

An invalid glob pattern was encountered.

InvalidRegexWithoutMessage(SourceSpan)

An invalid regex was encountered but we couldn’t determine a better error message.

ExpectedCloseRegex(SourceSpan)

A regex string was not closed.

InvalidOrOperator(SourceSpan)

An unexpected OR operator was found.

InvalidAndOperator(SourceSpan)

An unexpected AND operator was found.

UnexpectedArgument(SourceSpan)

An unexpected argument was found.

UnexpectedComma(SourceSpan)

An unexpected comma was found.

InvalidString(SourceSpan)

An invalid string was found.

ExpectedOpenParenthesis(SourceSpan)

An open parenthesis ( was expected but not found.

ExpectedCloseParenthesis(SourceSpan)

A close parenthesis ) was expected but not found.

InvalidEscapeCharacter(SourceSpan)

An invalid escape character was found.

ExpectedExpr(SourceSpan)

An expression was expected in this position but not found.

ExpectedEndOfExpression(SourceSpan)

The expression was expected to end here but some extra text was found.

NoPackageMatch(SourceSpan)

This matcher didn’t match any packages.

InvalidPlatformArgument(SourceSpan)

Expected “host” or “target” for a platform() predicate.

Unknown

An unknown parsing error occurred.